News_Chatbot/requirements.txt
python-dotenv
openai
requests
flask
flask-cors